+++ scanapi/__main__.py
@@ -37,6 +37,13 @@ def main():
     help="Run ScanAPI without generating report.",
 )
 @click.option(
+    "-b",
+    "--browser",
+    "open_browser",
+    is_flag=True,
+    help="Open the results file using a browser",
+)
+@click.option(
     "-c",
     "--config-path",
     "config_path",
@@ -58,7 +65,15 @@ def main():
     default="INFO",
     help="Set the debug logging level for the program.",
 )
-def run(spec_path, output_path, no_report, config_path, template, log_level):
+def run(
+    spec_path,
+    output_path,
+    no_report,
+    config_path,
+    template,
+    log_level,
+    open_browser,
+):
     """
     Automated Testing and Documentation for your REST API.
     SPEC_PATH argument is the API specification file path.
@@ -72,6 +87,7 @@ def run(spec_path, output_path, no_report, config_path, template, log_level):
         "no_report": no_report,
         "config_path": config_path,
         "template": template,
+        "open_browser": open_browser,
     }
 
     try: