+++ scanapi/settings.py
@@ -23,8 +23,9 @@ class Settings(dict):
         self["output_path"] = None
         self["template"] = None
         self["no_report"] = False
+        self["open_browser"] = False
 
-        super(dict, self).__init__()
+        super().__init__()
 
     def save_config_file_preferences(self, config_path=None):
         """Saves the Settings object config file preferences."""